+++ src/test/java/com/ironcorelabs/tenantsecurity/kms/v1/RekeyRoundTrip.java
@@ -50,18 +50,17 @@ public class RekeyRoundTrip {
         new DocumentMetadata(tenant_id, "integrationTest", "sample", customFields, "customRayID");
     Map<String, byte[]> documentMap = getRoundtripDataToEncrypt();
 
-    @SuppressWarnings("deprecation")
     CompletableFuture<PlaintextDocument> roundtrip =
         TenantSecurityClient.create(tsp_address + tsp_port, api_key).thenCompose(client -> {
           try {
             return client.encrypt(documentMap, metadata).thenCompose(encryptedDocument -> {
-              return client.rekeyDocument(encryptedDocument, metadata, new_tenant_id)
-                  .thenCompose(encryptedDocumentRekeyed -> {
-                    assertEquals(encryptedDocumentRekeyed.getEncryptedFields(),
-                        encryptedDocument.getEncryptedFields());
+              return client.rekeyEdek(encryptedDocument.getEdek(), metadata, new_tenant_id)
+                  .thenCompose(rekeyedEdek -> {
                     DocumentMetadata newMetadata = new DocumentMetadata(new_tenant_id,
                         "integrationTest", "sample", customFields, "customRayID");
-                    return client.decrypt(encryptedDocumentRekeyed, newMetadata);
+                    EncryptedDocument newDocument =
+                        new EncryptedDocument(encryptedDocument.getEncryptedFields(), rekeyedEdek);
+                    return client.decrypt(newDocument, newMetadata);
                   });
             });
           } catch (Exception e) {